refactor path construction and test double error
The logic needs improvements to identify which parts can be safely
removed.
As example, assume 123 are cases with 12 the errorneous cases.
There is no reasonable way to identify the source of double errors
without either 1. parse the error return trace output or
2. remove all tests but this test.

Option 2 is simpler, so we need to do this instead.