Reduction in test blocks require complex logic.
On the bright side, this will means a drive downhill except for
some annoyance of temporary fixups and the complexity to handle the
other parts.

Efficient and reusable fixups require to identify, which source
locations of the fixups between all test blocks are identical.
The main annoyance is that we dont have the comptime data and no
referenced by information yet, so we can only try.

On the bright side, the logic for long-term storage after long-term
committing to the fixups can be reused, but the bad side is that a
temporary storage is needed.

Thereafter the cool and interesting stuff will be done with package
resolvement and symbol index stuff.

Also, I need to nudge augustus into offering a custom command or
function to create a tags file and start prototyping into how to poke
Zig for the comptime info.